+++ tests/testthat/test_postgis_RPostgres.R
@@ -1,6 +1,17 @@
+#' To run the tests from the database, you can setup a docker container
+#' and run it as needed.
+#' docker run \
+#'   --name "postgis_test" \
+#'   -p 5432:5432 \
+#'   -e POSTGRES_USER=$USER \
+#'   -e POSTGRES_PASS=$USER \
+#'   -e POSTGRES_DBNAME=postgis \
+#'   -d -t kartoza/postgis
+#'  docker start postgis
+
 library(sf)
 library(DBI)
-library(RPostgreSQL)
+library(RPostgres)
 library(testthat)
 context("sf: postgis using RPostgres")
 
@@ -25,36 +36,40 @@ epsg_31370 = paste0("+proj=lcc +lat_1=51.16666723333333 +lat_2=49.8333339 ",
 
 pg <- NULL
 test_that("check utils", expect_false(can_con(pg)))
-try(pg <- DBI::dbConnect(RPostgres::Postgres(), host = "localhost", dbname = "postgis"), silent=TRUE)
+try(pg <- DBI::dbConnect(
+	RPostgres::Postgres(),
+	host = "localhost",
+	dbname = "postgis",
+	password = Sys.info()[["user"]]), silent=TRUE)
 
 # tests ------------------------------------------------------------------------
 test_that("can write to db", {
     skip_if_not(can_con(pg), "could not connect to postgis database")
     expect_silent(suppressMessages(st_write(pts, pg, "sf_meuse__")))
-    expect_error(st_write(pts, pg, "sf_meuse__"), "exists")
-    expect_silent(st_write(pts, pg, "sf_meuse__", overwrite = TRUE))
+    expect_error(st_write(pts, pg, "sf_meuse__", append = FALSE, delete_layer = FALSE), "exists")
+    expect_silent(st_write(pts, pg, "sf_meuse__", delete_layer = TRUE))
     expect_silent(st_write(pts, pg, "sf_meuse2__", binary = FALSE))
     expect_warning(z <- st_set_crs(pts, epsg_31370))
     expect_message(st_write(z, pg, "sf_meuse3__"), "Inserted local crs")
     expect_silent(st_write(z, pg, "sf_meuse3__", append = TRUE))
     expect_equal(nrow(DBI::dbReadTable(pg, "sf_meuse3__")), nrow(z) * 2)
-    expect_silent(st_write(z, pg, "sf_meuse3__", overwrite = TRUE))
+    expect_silent(st_write(z, pg, "sf_meuse3__", delete_layer = TRUE))
 })
 
 test_that("can handle multiple geom columns", {
     skip_if_not(can_con(pg), "could not connect to postgis database")
     multi <- cbind(pts[["geometry"]], st_transform(pts, 4326))
-    expect_silent(st_write(multi, pg, "meuse_multi", overwrite = TRUE))
+    expect_silent(st_write(multi, pg, "meuse_multi", delete_layer = TRUE))
     expect_silent(x <- st_read("PG:host=localhost dbname=postgis", "meuse_multi", quiet = TRUE))
     # expect_equal(st_crs(x[["geometry"]]), st_crs(multi[["geometry"]])) -> fails if EPSG databases differ
-    expect_equal(st_crs(x[["geometry.1"]]), st_crs(multi[["geometry.1"]]))
+    expect_true(st_crs(x[["geometry.1"]]) == st_crs(multi[["geometry.1"]]))
     expect_silent(x <- st_read("PG:host=localhost dbname=postgis", "meuse_multi", quiet = TRUE, type = c(1,4)))
     expect_silent(x <- st_read("PG:host=localhost dbname=postgis", "meuse_multi", quiet = TRUE, type = c(4,4)))
     expect_silent(x <- st_read("PG:host=localhost dbname=postgis", "meuse_multi", quiet = TRUE, promote_to_multi = FALSE))
     expect_silent(x <- st_read("PG:host=localhost dbname=postgis", "meuse_multi", quiet = TRUE, geometry_column = "geometry.1"))
     x <- st_layers("PG:host=localhost dbname=postgis")
     multi2 <- cbind(pts[["geometry"]], st_set_crs(st_transform(pts, 4326), NA))
-    expect_silent(st_write(multi2, pg, "meuse_multi2", overwrite = TRUE))
+    expect_silent(st_write(multi2, pg, "meuse_multi2", delete_layer = TRUE))
     expect_silent(x <- st_read(pg, "meuse_multi2"))
     expect_equal(st_crs(x[["geometry"]]), st_crs(multi2[["geometry"]]))
     expect_equal(st_crs(x[["geometry.1"]]), st_crs(multi2[["geometry.1"]]))
@@ -68,7 +83,7 @@ test_that("sf can write units to database (#264)", {
     ptsu <- pts
     ptsu[["u"]] <- ptsu[["cadmium"]]
     units(ptsu[["u"]]) <- units::as_units("km")
-    expect_silent(st_write(ptsu, pg, "sf_units__", overwrite = TRUE))
+    expect_silent(st_write(ptsu, pg, "sf_units__", delete_layer = TRUE))
     r <- st_read(pg, "sf_units__")
     expect_is(r[["u"]], "numeric")
     expect_equal(sort(r[["u"]]), sort(as.numeric(ptsu[["u"]])))
@@ -145,7 +160,7 @@ test_that("sf can preserve types (#592)", {
     # cannot write lists
     #dtypes$lst <- c(list(matrix("a")), list(matrix(c("b", "c"))), list(NA))
     dtypes <- st_as_sf(dtypes, coords = c("x", "y"))
-    st_write(dtypes, pg, overwrite = TRUE)
+    st_write(dtypes, pg, delete_layer = TRUE)
     x <- st_read(pg, "dtypes")
     dtypes$fact <- as.character(dtypes$fact)
     expect_equal(x, dtypes)
@@ -158,17 +173,19 @@ test_that("can write to other schema", {
     q <- "SELECT schema_name FROM information_schema.schemata WHERE schema_name = 'sf_test__';"
     suppressWarnings(could_schema <- DBI::dbGetQuery(pg, q) %>% nrow() > 0)
     skip_if_not(could_schema, "Could not create schema (might need to run 'GRANT CREATE ON DATABASE postgis TO <user>')")
-    expect_error(st_write(pts, pg, Id(schema = "public", table = "sf_meuse__")), "exists")
-    expect_silent(st_write(pts, pg, Id(schema = "sf_test__", table = "sf_meuse__")))
-    expect_error(st_write(pts, pg, Id(schema = "sf_test__", table = "sf_meuse__")), "exists")
-    expect_silent(st_write(pts, pg, Id(schema = "sf_test__", table = "sf_meuse__"), overwrite = TRUE))
+    tbl_meuse_public <- Id(schema = "public", table = "sf_meuse__")
+    tbl_meuse_test <- Id(schema = "sf_test__", table = "sf_meuse__")
+    expect_error(st_write(pts, pg, tbl_meuse_public, append = FALSE, delete_layer = FALSE), "exists")
+    expect_silent(st_write(pts, pg, tbl_meuse_test))
+    expect_error(st_write(pts, pg, tbl_meuse_test, append = FALSE, delete_layer = FALSE), "exists")
+    expect_silent(st_write(pts, pg, tbl_meuse_test, delete_layer = TRUE))
     expect_warning(z <- st_set_crs(pts, epsg_31370))
     expect_silent(st_write(z, pg, Id(schema = "sf_test__", table = "sf_meuse33__")))
     expect_silent(st_write(z, pg, Id(schema = "sf_test__", table = "sf_meuse4__")))
 
     # weird name work
-    expect_silent(st_write(pts, pg, c(NULL, "sf_test__.meuse__"), overwrite = TRUE))
-    expect_silent(st_write(pts.2 <- pts, pg, overwrite = TRUE))
+    expect_silent(st_write(pts, pg, c(NULL, "sf_test__.meuse__"), delete_layer = TRUE))
+    expect_silent(st_write(pts.2 <- pts, pg, delete_layer = TRUE))
     expect_true(DBI::dbRemoveTable(pg, "pts.2 <- pts"))
 })
 
@@ -328,54 +345,53 @@ test_that("Can safely manipulate crs", {
     skip_if_not(can_con(pg), "could not connect to postgis database")
     srid <- 4326
     crs <- st_crs(srid)
-    expect_true(get_postgis_crs(pg, srid) == st_crs(srid))
+    expect_true(find_database_srid(pg, srid = srid) == st_crs(srid))
+    expect_true(find_database_srtext(pg, crs) == st_crs(srid))
     expect_error(set_postgis_crs(pg, st_crs(srid)), "already exists")
     expect_warning(expect_true(is.na(st_crs(get_new_postgis_srid(pg)))), "not found")
-    new_crs <- st_crs(get_new_postgis_srid(pg), "+proj=longlat +datum=WGS84 +no_defs", valid = FALSE)
+    new_crs <- make_empty_crs(
+    	epsg = get_new_postgis_srid(pg),
+    	text = "+proj=longlat +datum=WGS84 +no_defs"
+    )
     expect_message(set_postgis_crs(pg, new_crs, auth_name = "sf_test"), "Inserted local crs")
-    expect_warning(expect_error(set_postgis_crs(pg, new_crs), "duplicate key"),
-                   "not found")
+    expect_error(set_postgis_crs(pg, new_crs), "duplicate key")
     expect_equal(delete_postgis_crs(pg, new_crs), 1)
     expect_equal(delete_postgis_crs(pg, new_crs), 0)
 
     # set and delete
-    crs$epsg <- NA
-    expect_message(new_srid <- set_postgis_crs(pg, crs), "Inserted local crs")
-    expect_error(delete_postgis_crs(pg, crs), "Missing SRID")
-    crs2 <- st_crs(new_srid$epsg, proj4text = st_crs(3857)$proj4string, valid = FALSE)
+    new_crs <- make_empty_crs(
+    	epsg = NA,
+    	text = st_as_text(st_crs(4326))
+    )
+    expect_message(new_srid <- set_postgis_crs(pg, new_crs), "Inserted local crs")
+    expect_error(delete_postgis_crs(pg, new_crs), "Missing SRID")
+
+    crs2 <- make_empty_crs(epsg(new_srid), st_as_text(st_crs(3857)))
     expect_equal(delete_postgis_crs(pg, crs2), 0)  # crs doesn't match any crs
     expect_equal(delete_postgis_crs(pg, new_srid), 1)
 
     # udpate
     expect_message(set_postgis_crs(pg, new_srid), "Inserted local crs")
-    new_srid$proj4string <- crs2$proj4string
-    expect_warning(
-        expect_error(set_postgis_crs(pg, new_srid), "already exists"),
-        "GDAL Error 6: EPSG"
-    )
-    expect_warning(
-        expect_message(set_postgis_crs(pg, new_srid, update = TRUE), "Inserted local crs"),
-        "GDAL Error 6: EPSG"
-    )
-
+    new_srid[["wkt"]] <- crs2[["wkt"]]
+    expect_error(set_postgis_crs(pg, new_srid), "already exists")
+    expect_message(set_postgis_crs(pg, new_srid, update = TRUE), "Inserted local crs")
 })
 
 
 test_that("new SRIDs are handled correctly", {
     skip_if_not(can_con(pg), "could not connect to postgis database")
-    data(meuse, package = "sp")
-    meuse_sf = st_as_sf(meuse, coords = c("x", "y"), crs = NA_crs_)
+	skip("TODO: Need to produce a locally unknown crs")
+	crs <- make_empty_crs(
+		epsg = NA,
+		wkt = "locally unknown crs"
+	)
+    st_crs(pts) <- crs
 
-    crs = st_crs(NA_integer_, paste("+proj=sterea +lat_0=52 +lon_0=5", # creates FALSE, but new one
-                                    "+k=1.0 +x_0=155000 +y_0=463000 +ellps=bessel",
-                                    "+towgs84=565.4171,50.3319,465.5524,-0.398957,0.343988,",
-                                    "-1.87740,4.0725 +units=m +no_defs"), valid = FALSE)
-    st_crs(meuse_sf) = crs
-    expect_message(st_write(meuse_sf, pg, overwrite = TRUE), "Inserted local crs")
+    expect_message(st_write(pts, pg, "meuse_sf", delete_layer = TRUE), "Inserted local crs")
     expect_warning(x <- st_read(pg, query = "select * from meuse_sf limit 3;"),
                    "not found in EPSG support files")
     expect_true(st_crs(x)$proj4string == crs$proj4string)
-    expect_silent(st_write(meuse_sf, pg, overwrite = TRUE))
+    expect_silent(st_write(meuse_sf, pg, delete_layer = TRUE))
 })
 
 test_that("schema_table", {